api_basic/serializers/article.py
from rest_framework import serializers
from api_basic.models import Article


class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = '__all__'

