requirements.txt
djangorestframework==3.12.4
mysqlclient==2.0.3
python-dotenv==0.19.1
Django~=4.0.4